refactor: move logic into src modules
